Create Node.java
This class stores the puzzle state, parent node, action taken to get from p to the current node, cost so far, estimated cost to get to the goal, total cost (cost so far + estimated cost), and depth. This particular node class is designed to work for A* Searches, since the total cost is estimated from the addition of the cost so far and estimated cost.