Create TestCase.java
This class defines the TestCase object, which stores a solution Node that results from an A* Search, total number of nodes generated to find that result, the depth at which that node was, whether or not the first or second heuristic function was used to find that solution, and the amount of time taken to find that solution. This class is used when displaying the solution analytics of multiple solutions.